backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests

pypdf
sentence-transformers
numpy
psycopg2-binary
